Revert "LibWeb: Make TextNode::ChunkIterator emit an empty chunk for content:"""
This reverts commit 360f1ed1f9a65d9182b9523026a2c9292d1a4e76.

This made a calculation of pseudo-elements' height incorrect when they
had `height` set to `auto` and used other techniques (like setting
`padding-top`) to set height, as it was now also adding an empty line.

Additionally, the case didn't work for content containing whitespace
characters, so a pseudo-element with `content: " "` didn't have *this*
particular problem.